packages/cosma-backend/pyproject.toml
[project]
name = "cosma-backend"
version = "0.1.1"
description = "Cosma backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asqlite>=2.0.0",
    "litellm>=1.77.5",
    "llama-cpp-python",
    "markitdown[docx,pdf,pptx]>=0.1.3",
    "ollama>=0.6.0",
    "quart>=0.20.0",
    "quart-schema>=0.22.0",
    "rich>=14.2.0",
    "sentence-transformers>=5.1.1",
    "sqlite-vec>=0.1.6",
    "tiktoken>=0.11.0",
    "uvicorn>=0.38.0",
    "watchdog>=6.0.0",
]

[tool.pyright]
typeCheckingMode = "standard"

[dependency-groups]
dev = [
    "basedpyright<=1.31.7",
    "pyinstaller>=6.16.0",
]
test = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.0",
    "pytest-cov>=3.15.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml"
]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (database, file system)",
    "slow: Slow tests (network calls, AI services)",
    "async_test: Tests that use async/await"
]

[tool.uv.sources]
llama-cpp-python = { git = "https://github.com/JamePeng/llama-cpp-python" }

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
